backend/requirements.txt
# ── Web framework ──────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Database ───────────────────────────────────────────
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# ── Cache / Queue ──────────────────────────────────────
redis==5.0.3
celery==5.3.6
kombu==5.3.4

# ── Auth ───────────────────────────────────────────────
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# ── Data processing ────────────────────────────────────
pandas==2.2.1
numpy==1.26.4
openpyxl==3.1.2
xlrd==2.0.1
scipy==1.13.0
scikit-learn==1.4.2
statsmodels==0.14.1
pyarrow==15.0.2

# ── AI / LLM ───────────────────────────────────────────
anthropic==0.25.1
openai==1.25.0
tiktoken==0.6.0

# ── File / Storage ─────────────────────────────────────
boto3==1.34.84
python-magic==0.4.27
Pillow==10.3.0

# ── Report generation ──────────────────────────────────
reportlab==4.1.0
python-pptx==0.6.23
jinja2==3.1.3
weasyprint==62.3

# ── Validation ─────────────────────────────────────────
pydantic==2.7.0
pydantic-settings==2.2.1
email-validator==2.1.1

# ── Utils ──────────────────────────────────────────────
httpx==0.27.0
python-dotenv==1.0.1
loguru==0.7.2
tenacity==8.2.3

# ── Testing ────────────────────────────────────────────
pytest==8.1.1
pytest-asyncio==0.23.6
httpx==0.27.0
factory-boy==3.3.0
